+++ tmp/swagger-test-3x7yud/product-types.ts
@@ -1,9 +1,0 @@
-/**
- * Auto-generated types from Swagger/OpenAPI
- * Generated at 2026-02-20T20:51:48.509Z
- */
-
-export interface Product {
-  price?: number;
-}
-